combind ground intake
