+++ src/main/java/org/usfirst/frc4904/robot/subsystems/Manipulator.java
@@ -4,12 +4,9 @@ import org.usfirst.frc4904.standard.subsystems.SolenoidSubsystem;
 import org.usfirst.frc4904.standard.subsystems.motor.Motor;
 
 public class Manipulator {
-	public static final double INTAKE_SPEED = -1; //TODO: change these
-	public static final double OUTTAKE_SPEED = -1;
-	public static final double KEEP_BALL_SPEED = -1;
-	public static final double INTAKE_TIME = -1;
-	public static final double OUTTAKE_TIME = -1;
-	public static final double STOP_TIME = -1;
+	public static final double INTAKE_SPEED = 0.7; //TODO: change these
+	public static final double OUTTAKE_SPEED = -0.7;
+	public static final double KEEP_BALL_SPEED = -0.1;
 	public final SolenoidSubsystem wrist; // manipulator arm
 	public final SolenoidSubsystem claws; // cargo skid
 	public final SolenoidSubsystem hatchExtender; // the ground intake
